[REFACTOR] Physics no longer requires camera.h
