+++ src/world/mapbg.c
@@ -1,4 +1,4 @@
-#include <FE_Client.h>
+#include "../client/core/include/include.h"
 
 #include "include/include.h"
